+++ src/constants/ports.py
@@ -1,0 +1,11 @@
+SSL_PORT = 443
+HTTP_PORT = 80
+SQL_PORT = 3306
+SMTP_PORT = 3306
+
+ports = {
+  "ssl": SSL_PORT,
+  "http": HTTP_PORT,
+  "sql": SQL_PORT,
+  "mail": SMTP_PORT,
+}