+++ CHANGELOG.md
@@ -1,8 +1,19 @@
 version: 1.1.0
 changelog:
-1. DDOS basic (active)
+1. DDOS BASIC (active)
 2. DDOS ALL REQUEST (active)
 3. IP REVERSE (active)
 4. TEST PROXIES (active)
 5. UPDATE PROXIES (active)
 6. SPAM POST (active)
+
+version: 1.2.0
+changelog:
+1. Add custom max_workers at config.json file
+2. Add more data for SPAM POST menu
+
+version: 1.2.5
+changelog:
+1. Add port in every request for intensive layer
+2. SMTP KILLER (active)
+2. DNS RECORD (active)